+++ relay-server/src/services/processor.rs
@@ -99,7 +99,6 @@ mod event;
 mod metrics;
 mod nel;
 mod profile;
-pub(crate) mod replay;
 mod report;
 mod span;
 
@@ -564,12 +563,6 @@ pub enum ProcessingError {
     #[error("invalid processing group type")]
     InvalidProcessingGroup(Box<InvalidProcessingGroupType>),
 
-    #[error("invalid replay")]
-    InvalidReplay(DiscardReason),
-
-    #[error("replay filtered with reason: {0:?}")]
-    ReplayFiltered(FilterStatKey),
-
     #[cfg(feature = "processing")]
     #[error("nintendo switch dying message processing failed {0:?}")]
     InvalidNintendoDyingMessage(#[source] SwitchProcessingError),
@@ -621,8 +614,6 @@ impl ProcessingError {
             Self::MissingProjectId => None,
             Self::EventFiltered(_) => None,
             Self::InvalidProcessingGroup(_) => None,
-            Self::InvalidReplay(reason) => Some(Outcome::Invalid(*reason)),
-            Self::ReplayFiltered(key) => Some(Outcome::Filtered(key.clone())),
 
             Self::ProcessingGroupMismatch => Some(Outcome::Invalid(DiscardReason::Internal)),
             // Outcomes are emitted in the new processing pipeline already.
@@ -1757,40 +1748,6 @@ impl EnvelopeProcessorService {
             self.inner.addrs.outcome_aggregator.clone(),
         );
 
-        Ok(Some(extracted_metrics))
-    }
-
-    // TODO: Start porting this logic over
-    /// Processes replays.
-    async fn process_replays(
-        &self,
-        managed_envelope: &mut TypedEnvelope<ReplayGroup>,
-        ctx: processing::Context<'_>,
-    ) -> Result<Option<ProcessingExtractedMetrics>, ProcessingError> {
-        // Q: Check how this is handled in the other ports
-        let mut extracted_metrics = ProcessingExtractedMetrics::new();
-
-        // META: Do the processing
-        // Q: Is this taking a managed envelop an issue?
-        replay::process(
-            managed_envelope,
-            ctx.global_config,
-            ctx.config,
-            ctx.project_info,
-            &self.inner.geoip_lookup,
-        )?;
-
-        // META: Enforce the quotas
-        // Q: What is being extracted that needs the checks here.
-        self.enforce_quotas(
-            managed_envelope,
-            Annotated::empty(),
-            &mut extracted_metrics,
-            ctx,
-        )
-        .await?;
-
-        // Obs: This seems to be the output of the processing
         Ok(Some(extracted_metrics))
     }
